+++ js/map.js
@@ -5,7 +5,7 @@ const MAIN_PIN_WIDTH_LENGTH = [52, 52];
 const MAIN_PIN_ANCHOR_XY = [26, 52];
 const PIN_WIDTH_LENGTH = [40, 40];
 const PIN_ANCHOR_XY = [20, 40];
-
+const address = document.querySelector('#address');
 const map = initMap();
 
 const mainPinIcon = L.icon({
@@ -37,15 +37,12 @@ const errorMessageMarkers = (err) => {
   setInterval(toggleError, 2000);
   mapCanvas.append(div);
 };
-let mainMarker;
+
 const creatMainMarker = (currentMap, currentLatLng, icon) => {
-  const address = document.querySelector('#address');
+
   address.setAttribute('readonly', 'readonly');
-  const handlerMarkerOnMoveend = (evt) => {
-    const {lng, lat} = evt.target.getLatLng();
-    address.value = `${lng.toFixed(5)},${lat.toFixed(5)}`;
-  };
-  mainMarker = L.marker(
+
+  const mainMarker = L.marker(
     currentLatLng,
     {
       draggable: true,
@@ -54,8 +51,7 @@ const creatMainMarker = (currentMap, currentLatLng, icon) => {
   );
   document.querySelector('#address').value = Object.values(currentLatLng);
 
-  mainMarker.on('moveend', handlerMarkerOnMoveend).addTo(currentMap);
-
+  return mainMarker;
 };
 
 
@@ -72,12 +68,29 @@ const creatMarkers = (currentMap, mainIcon, ads) => {
   ads.forEach((ad) => {
     creatMarker(ad);
   });
+  return markerGroup;
 };
 
 const latLng = map.getCenter();
+const zoom = map.STARING_ZOOM;
 const getCenterMap = () => map.getCenter();
 const creatMainMarkerOnMap = () => creatMainMarker(map, latLng , mainPinIcon);
 const creatMarkersOnMap = (ads) => creatMarkers(map, pinIcon, ads);
 
-export {creatMainMarkerOnMap, creatMarkersOnMap, errorMessageMarkers, getCenterMap};
+
+const mainMarkerCurrent = creatMainMarker(map, latLng , mainPinIcon);
+mainMarkerCurrent.addTo(map);
+const handlerMarkerOnMoveend = (evt) => {
+  const {lng, lat} = evt.target.getLatLng();
+  address.value = `${lng.toFixed(5)},${lat.toFixed(5)}`;
+};
+mainMarkerCurrent.on('moveend', handlerMarkerOnMoveend).addTo(map);
+const resetMainPinMarker = () => {
+  map.closePopup();
+  map.setView(new L.LatLng(latLng.lat, latLng.lng), zoom);
+  mainMarkerCurrent.setLatLng(L.latLng(latLng));
+
+};
+
+export {creatMainMarkerOnMap, creatMarkersOnMap, errorMessageMarkers, getCenterMap, resetMainPinMarker};
